only call set_texts when we need to
Since set_texts resets self.first_active_idx to 0 (otherwise, you could
have it be an index that's larger than the new number of texts!), we
need to only call it when we actually have new words. If we call it at
every loop (as we were), then scrolling up/down gets overwritten,
and we're stuck seeing just the first page of results.