adding a new analyzer
This one measures each char not by how well it bisects the word list,
but by how well it divides it into three. To do this, it takes the char
positions into account. At each position, a char can either be (a)
completely right, in the wrong position, or completely wrong: three
options. The closer we are to even, the better.

Consider three words:

    ACE
    DAB
    MOB

Guessing "Z" for position zero doesn't help at all: all three options
are still valid. Guessing "B" is a little better: if the answer is "ACE"
then you'll know it (since you'll be told there's no "B", and "ACE" is
the only word without one), but if not, "DAB" or "MOB" are both equally
valid. But if you guess "A", you'll be guaranteed to know it: among the
words, "A"-in-pos0 is distributed evenly: there's one word where it's
the correct char, one word where it's in the wrong position, and one
word where it's missing altogether. That means that if you guess a word
with "A at the start, you're guaranteed to get the solution on your
second try.

Obviously, the actual words list isn't that simple, and especially there
are correlations between char-positions. One place this especially falls
is in words that repeat a char: this algorithm likes those (since the
char is likely to be good in multiple positions), but you get
diminishing returns from each one -- you only gain the "missing
altogether" datapoint from one of them.

But, overall the ope is that if each char in a guess roughly divides the
list into 3, then with 5 guesses we'll have pretty good segmentation.