handle unknown input better
I forget why this was a loop at all, but that's broken. It means that if
none of the input handlers know how to deal with a given input, then it
stays as a `Some(_)`, and thus loops back to the top.

I think the outer loop (which this commit removes) was my first attempt
at the input handlers loop. Since that's now done inside, I was
basically double-looping.