pre-fill any known chars
Pre-fill them, but don't skip over them when typing. Otherwise, it mucks
with muscle memory. If I have the word "HELLO" in my head, it's easier
to type "HELLO" then to remember that the first "L" is pre-filled, so I
just need to type "HELO".

If you overwrite a known char with something else, its knowledge becomes
Unknown. If you then re-type the known letter, it'll become Correct
automatically.

This resolves #2.