warn if a partially-typed word isn't possible
This is working, but I don't like it. Let's say you had typed:

    ABOUT

and then told the app, "ABOU" are wrong, but T is there but in the wrong
position." When you hit <enter>, it'll take you to the next line and
keep it blank -- and instantly complain that the word so far (which is
blank) is already invalid, because it doesn't contain any "T".

I need to be a bit less precise with this, and probably only judge
whether the letters I _have_ typed are impossible. And maybe go into
"strict" mode when you fill all the letters out?

idk, I need to think about it a bit.

This is in service of #3.